LPS-206029 Extract to reuse, no logic changes
